Fix issue where switching scenes would cause reconnects
This would happen regardless of whether the device or plugin were in the next scene. If there was an iOS Camera input that was duplicated in two scenes, then the plugin would disconnect and reconnect during a scene switch. This occasionally caused issues where the device would fail to reconnect.